add ProcessMessages() in main loop //Обработка различных флагов и сообщений от задач

+++ RTOS/EERTOS.h
@@ -14,7 +14,7 @@ typedef void (*TPTR)(void);
 
 #ifdef DEBUG
  extern  volatile uint32_t v_u32_SYS_TICK;    
- extern  volatile uint32_t v_u32_SYS_TICK_TMP1;
+ extern  volatile uint8_t v_u32_SYS_TICK_TMP1;
  extern void Put_In_Log (unsigned char * data);
  
  extern volatile uint16_t v_u16_TIM_1_OVR_FLAG;
